+++ src/main/java/ICreditCalc.java
@@ -1,9 +1,7 @@
-import java.math.BigDecimal;
-
 public interface ICreditCalc {
-    public BigDecimal monthlyPayment(Credit credit);
+    public int monthlyPayment(Credit credit);
 
-    public BigDecimal totalAmountPlusPercent(Credit credit);
+    public double totalAmountPlusPercent(Credit credit);
 
-    public BigDecimal overpayment(Credit credit);
+    public double overpayment(Credit credit);
 }
